Fix bug caused by adding new lines in gem definitions
This fix brings pessimize to the point where practically all
Gemfiles will correctly parsed, and therefore closes #5.

The fix was to allow for the :on_ignored_nl token to be recognised when
searching for the version strings in the list of gem tokens.

This commit also adds supporting specs and integration tests.